pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG system with OpenWebUI, CrewAI, LlamaIndex, and Arize Phoenix."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "Candidate" }]
dependencies = [
    # --- Core API & Async ---
    "fastapi>=0.128.0,<1.0.0",
    "uvicorn[standard]>=0.40.0,<1.0.0",
    "python-multipart>=0.0.22,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "tenacity>=9.1.0,<10.0.0",

    # --- Data & Settings ---
    "pydantic>=2.11.0,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "pyyaml>=6.0.3,<7.0.0",

    # --- Database & Vector ---
    "sqlalchemy>=2.0.46,<3.0.0",
    "asyncpg>=0.31.0,<1.0.0",
    "pgvector>=0.4.0,<1.0.0",

    # --- AI & RAG Logic ---
    "llama-index-core>=0.14.0,<0.15.0",
    "llama-index-llms-ollama>=0.9.0,<0.10.0",
    "llama-index-embeddings-ollama>=0.8.0,<0.9.0",
    "llama-index-vector-stores-postgres>=0.7.0,<0.8.0",
    "llama-index-readers-docling>=0.4.0,<0.5.0",
    "docling>=2.70.0,<3.0.0",
    "crewai>=1.9.0,<2.0.0",
    "litellm>=1.75.0,<2.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    "ollama>=0.6.0,<1.0.0",

    # --- Observability ---
    "arize-phoenix>=12.33.0,<13.0.0",
    "arize-phoenix-client>=1.0.0,<2.0.0",
    "openinference-instrumentation-llama-index>=4.3.0,<5.0.0",
    "openinference-instrumentation-crewai>=0.1.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "openinference-semantic-conventions>=0.1.25,<1.0.0",

    # --- Logging & CLI ---
    "structlog>=25.0.0,<30.0.0",
    "typer[all]>=0.21.0,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "jinja2>=3.1.6,<4.0.0",
]

[project.optional-dependencies]
hf = [
    "transformers>=4.40.0",
]
eval = [
    "ragas>=0.4.3,<1.0.0",
    "pandas>=2.1.4,<4.0.0",
    "datasets>=4.5.0,<5.0.0",
]
dev = [
    "ruff>=0.14.0,<1.0.0",
    "mypy>=1.19.0,<2.0.0",
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "types-pyyaml>=6.0.12,<7.0.0",
]

[project.scripts]
agentic-index = "agentic_rag.indexer.cli:app"
agentic-eval = "agentic_rag.evaluator.cli:app"
agentic-api = "agentic_rag.backend.main:start"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agentic_rag = ["prompts/*.j2", "prompts/*.txt"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203", "B008"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = false
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "asyncpg",
    "pgvector.*",
    "openinference.*",
    "ragas.*",
    "datasets",
    "pandas",
]
ignore_missing_imports = true
follow_untyped_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
